+++ .gitignore
@@ -4,4 +4,3 @@ pan modeling/__pycache__/
 pan modeling/data/
 pan modeling/scripts/
 pan modeling/checkpoints/
-2_report
