+++ 1_모델링/UNet/.gitignore
@@ -1,0 +1,16 @@
+*.pyc
+data/
+__pycache__/
+checkpoints/
+*.jpg
+venv/
+.idea/
+.github/
+scripts/
+Dockerfile
+*.png
+wandb/
+test_modeling_any.ipynb
+requirements.txt
+hubconf.py
+train_testver.py
