+++ 1_모델링/UNet/INTERRUPTED.pth
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ec9f1079300aaf03336340f531aa0fbaad5127e21eb643dd952d9023e76e59c3
+size 69126693
