+++ 1_모델링/UNet/weight/UNet_b1th10dn10k.pth
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:60290aeff50afe929d818e12927e0f486d6dcd57fff6a5f4d3bec697bfa0567d
+size 124239509
