+++ 1_모델링/UNet/weight/UNet_b2th10dn50k.pth
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:25845594cd5a1b450ae1d06725d57c74182b99c5171c44b24814c195a7a1ead9
+size 124239509
