+++ 1_모델링/UNet/weight/UNet_b2th5dn10k.pth
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:4ad160a98d4199e0d85fe1be12cd5d3cdc59834e084afcd341ccf88e9276cb7b
+size 124239509
