+++ 1_모델링/UNet/weight/UNet_b2th5dn200k.pth
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0d54b84d14b80fd11fb264405fd46325708d84b8e59f0480a27e24de7de7ae08
+size 124239509
