--- 1_모델링/test_unet/.gitignore
@@ -1,20 +1,0 @@
-*.pyc
-data/
-__pycache__/
-checkpoints/
-*.pth
-*.jpg
-venv/
-.idea/
-.github/
-scripts/
-Dockerfile
-*.png
-wandb/
-test_modeling_any.ipynb
-requirements.txt
-hubconf.py
-<<<<<<< HEAD
-=======
-train_testver.py
->>>>>>> b21dbb31461ff9a9454c5184dcdb918fd063a3c7
