--- 1_모델링/test_unet/unet/__init__.py
@@ -1,1 +1,0 @@
-from .unet_model import UNet
